fix: Lỗi thư viện datatable export và sửa công thức tính tổng chi phí LL tại controller

+++ src/main/java/com/mobifone/transmission/controller/HomeController.java
@@ -64,9 +64,9 @@ public class HomeController {
         return totalFoSites;
     }
     @ModelAttribute("totalLeaseLineCostPerMonth")
-    public int getTotalLeaseLineCostPerMonth(){
+    public float getTotalLeaseLineCostPerMonth(){
         List<LeaseLine> leaseLines = leaseLineService.findAll();
-        int cost =0 ;
+        float cost = 0f ;
         for (LeaseLine ll : leaseLines) {
             cost += ll.getCost()*ll.getQuantity();
         }
